requirements.txt
# Main package
pipecat-ai

# Development dependencies
build~=1.2.2
coverage~=7.6.12
grpcio-tools~=1.67.1
pip-tools~=7.4.1
pre-commit~=4.0.1
pyright~=1.1.397
pytest~=8.3.4
pytest-asyncio~=0.25.3
pytest-aiohttp==1.1.0
ruff~=0.11.1
setuptools~=70.0.0
setuptools_scm~=8.1.0

# Application dependencies
python-dotenv~=1.0.1
fastapi~=0.110.0 
twilio
uvicorn~=0.27.1
onnxruntime
websockets
python-multipart
